Move types and functions into modules
The `main` executable includes all of the implementation of the failover
controller.

This change moves most of the failover controller's business logic into
modules to facilitate testing and to start making the `main` executable
focused more on configuration & initialization.
